Collect deprecated resource into single file
